separate lvgl widget func
